Wellbeing Coach/pyproject.toml
[project]
name = "wellbeing-coach"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "openai>=2.26.0",
    "psycopg2>=2.9.11",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "slowapi>=0.1.9",
    "uvicorn>=0.41.0",
]
